depare: generalize nodata to stem resolution; contours share the deep-coarsened DEM
Nodata rows carried full-detail OSM water outlines regardless of the stem's
DEM resolution — ~3/4 of the 20 GB+ planet depare archive, and under the
variable-depth pyramid they kept coarse inland stems subdividing to z13/14
instead of leafing (measured -54% at 1 px / -67% at 2 px on a cz8 stem).
Simplify each post-clip nodata piece to NODATA_SIMPLIFY_PX (default 1) times
the child_z resolution; post-clip so simplification can never push a boundary
across the seam. Bands and drying stay raw.

Move deep_coarsen from depare_run to smooth.py and call it from contour_run's
tile() as well: deep isobaths and depth-area band edges below -250 m now
derive from the same coarsened surface instead of visibly disagreeing.
soundings_run still reads the full-res window (unchanged on purpose).
