+++ pipelines/depare_run.py
@@ -61,6 +61,7 @@ import rasterio
 import config
 import contour_run
 import utils
+from aggregation_reproject import get_resolution
 
 # The bands' zoom floor (z6, matching the style's depth-areas/contour-lines minzoom) is now applied
 # per feature by the vector bundle (contour_run.DEPARE_MINZOOM): partitions can't be level-thinned
@@ -96,24 +97,19 @@ DRYING_RANK = 2
 # a width/compactness gate is the targeted tool if thin ribbons ever appear. Env-tunable.
 SLIVER_MIN_PX = float(os.environ.get("SLIVER_MIN_PX", "4"))
 
+# Nodata outlines are full-detail OSM geometry regardless of the stem's real resolution, so under
+# the variable-depth pyramid coarse inland tiles would keep subdividing to carry vertices no coarse
+# stem can resolve. Generalizing nodata rows to the stem's child_z resolution (this many MVT pixels)
+# is what lets those tiles leaf early (measured −54% at 1 px / −67% at 2 px on a cz8 stem; 21.5 M →
+# 1.4 M vertices, sub-resolution features dropping out naturally). Bands and drying are left raw —
+# their shared edges must stay bit-identical for the crack-free partition.
+NODATA_SIMPLIFY_PX = float(os.environ.get("NODATA_SIMPLIFY_PX", "1"))
+
 # Fixed-precision grid (metres) for overlays against multi-piece unions: GEOS 3.13's
 # float OverlayNG returns an empty overlay against some unions whose pairwise overlays are
 # correct (verified on 6-21-22-9; point-in-polygon arbitration); snap-rounded overlay is the
 # robust mode, and 1 µm moves no vertex cartographically.
 GRID = 1e-6
-
-# Deep-water DEM coarsening: below the threshold, band edges carry no navigational detail, but
-# source noise at fine child_z resolution makes gdal_contour -p's rings astronomically complex
-# (the deep-basin z14 stems burned 8.5 h+ of pure CPU; docs/plans/2026-07-21-depare-perf.md).
-# Pixels whose 8x-block mean is deeper than the threshold take that mean, so the deep interior
-# contours at ~8x resolution while everything shoaler keeps full detail. -250 m sits between
-# ladder levels (m: -200/-300; ft: -182.88/-365.76), so no level rides the transition. Only
-# fine windows (child_z >= 12) coarsen — coarse windows are already at or past 8x.
-DEEP_COARSEN_THRESHOLD_M = -250.0
-DEEP_COARSEN_FACTOR = 8
-DEEP_COARSEN_MIN_CHILD_Z = 12
-MERC_ORIGIN = 20037508.342789244  # EPSG:3857 half-extent; block grid anchors here (seam alignment)
-
 
 class ContourTimeout(Exception):
     """A bounded gdal_contour invocation exceeded DEPARE_TIMEOUT."""
@@ -129,40 +125,6 @@ def _run_bounded(cmd, what, timeout):
         if "(exit 124)" in str(e) or "(exit 137)" in str(e):
             raise ContourTimeout(f"{what} exceeded {timeout}s") from e
         raise
-
-
-def _deep_coarsen(dem, threshold=DEEP_COARSEN_THRESHOLD_M, factor=DEEP_COARSEN_FACTOR):
-    """Rewrite `dem` in place with deep pixels replaced by their block mean. The block grid is
-    anchored to the EPSG:3857 origin, NOT the window, so overlapping windows (neighbor stems'
-    buffers) average identical blocks and band edges still match at macrotile seams. A pixel is
-    replaced only when it AND its block mean are below the threshold: shelf-edge blocks whose
-    mean dips deep keep their shallow pixels, and a deep canyon inside a shallow block stays
-    fine rather than taking a shallow mean. Nodata pixels are never touched (the nodata pass
-    reads this same raster), and nodata is excluded from block means."""
-    import numpy as np
-    with rasterio.open(dem, "r+") as d:
-        arr = d.read(1)
-        nd = d.nodata
-        res = abs(d.transform.a)
-        h, w = arr.shape
-        # Global block alignment: pad to the origin-anchored grid (pads land in the buffer).
-        col0 = round((d.transform.c + MERC_ORIGIN) / res)
-        row0 = round((MERC_ORIGIN - d.transform.f) / res)
-        padt, padl = row0 % factor, col0 % factor
-        padb, padr = (-(h + padt)) % factor, (-(w + padl)) % factor
-        p = np.pad(arr, ((padt, padb), (padl, padr)), mode="edge")
-        valid = p != nd
-        ph, pw = p.shape
-        vals = np.where(valid, p, 0.0)
-        bsum = vals.reshape(ph // factor, factor, pw // factor, factor).sum(axis=(1, 3))
-        bcnt = valid.reshape(ph // factor, factor, pw // factor, factor).sum(axis=(1, 3))
-        del p, vals, valid
-        bmean = (bsum / np.maximum(bcnt, 1)).astype(np.float32)
-        bmean[bcnt == 0] = nd
-        up = np.repeat(np.repeat(bmean, factor, 0), factor, 1)[padt:padt + h, padl:padl + w]
-        mask = (arr != nd) & (arr <= threshold) & (up != nd) & (up <= threshold)
-        arr[mask] = up[mask]
-        d.write(arr, 1)
 
 
 def _uniform_coarsen(dem, factor, out):
@@ -254,7 +216,7 @@ def partitions(dem, levels, raw_fgb, timeout=0):
     return g
 
 
-def _depare_dem(dem, tile_obj, tmp, label, timeout=0):
+def _depare_dem(dem, tile_obj, child_z, tmp, label, timeout=0):
     """Partition any DEM covering the tile's buffered extent into depth-area / drying / nodata
     rows. Returns (final_path, count) inside ``tmp``, or None when there is no water."""
     import geopandas as gpd
@@ -270,6 +232,7 @@ def _depare_dem(dem, tile_obj, tmp, label, timeout=0):
     bbox = (b.left, b.bottom, b.right, b.top)  # the DEM's full (buffered) extent, EPSG:3857
     buffered = box(*bbox)
     min_area = SLIVER_MIN_PX * res * res       # slivers where a vector edge meets the raster shore
+    nodata_tol = NODATA_SIMPLIFY_PX * get_resolution(child_z)  # generalize nodata to the stem's resolution
     rows = []
 
     # ── depth bands + drying ── the metre + fathom partition ladders, each off one gdal_contour -p
@@ -396,8 +359,16 @@ def _depare_dem(dem, tile_obj, tmp, label, timeout=0):
             for full in _polys(geom):  # gate the PRE-clip polygon (buffered window) so a seam sliver survives both sides
                 if full.area >= min_area:
                     for p in _polys(full.intersection(clip)):  # then clip to the seam; no re-filter on the piece
-                        rows.append({"geometry": p, "drval1": None, "drval2": None,
-                                     "sys": None, "kind": kind, "rank": NODATA_RANK})
+                        # Simplify POST-clip, per-piece: kept vertices are a subset inside the clip
+                        # box, so the ring can never cross the seam outward — at worst it recedes
+                        # ≤ tol from the clip line (resolution-scale wobble, like the raster staircase).
+                        # Pre-clip would let far-away geometry change vertex picks near the seam.
+                        s = p.simplify(nodata_tol, preserve_topology=True)
+                        if s.is_empty:
+                            continue
+                        for sp in _polys(s):
+                            rows.append({"geometry": sp, "drval1": None, "drval2": None,
+                                         "sys": None, "kind": kind, "rank": NODATA_RANK})
         _mark("nodata-loop")
 
     if not rows:
@@ -446,17 +417,17 @@ def tile(stem):
         if not os.environ.get("SKIP_SMOOTH"):
             smooth.smooth_tiff(dem)
             _mark("smooth")
-        if child_z >= DEEP_COARSEN_MIN_CHILD_Z:
-            _deep_coarsen(dem)
+        if child_z >= smooth.DEEP_COARSEN_MIN_CHILD_Z:
+            smooth.deep_coarsen(dem)
             _mark("deep-coarsen")
         tile_obj = mercantile.Tile(x=x, y=y, z=z)
         try:
-            res = _depare_dem(dem, tile_obj, tmp, stem, timeout=timeout)
+            res = _depare_dem(dem, tile_obj, child_z, tmp, stem, timeout=timeout)
         except ContourTimeout as e:
             print(f"depare tile {stem}: {e} — retrying on a uniform 4x-average window",
                   file=sys.stderr, flush=True)
             dem = _uniform_coarsen(dem, 4, f"{tmp}/dem-4x.tiff")
-            res = _depare_dem(dem, tile_obj, tmp, stem, timeout=timeout)
+            res = _depare_dem(dem, tile_obj, child_z, tmp, stem, timeout=timeout)
         os.makedirs(os.path.dirname(out), exist_ok=True)
         if res:
             final, n = res
@@ -508,42 +479,29 @@ def _check():
     except ContourTimeout:
         pass
 
-    # _deep_coarsen: shallow and nodata pixels preserved; deep pixels take their block mean;
-    # and the block grid is origin-anchored, so two windows offset against the block grid
-    # produce identical values where they overlap (the macrotile-seam contract).
-    def _write_dem(path, arr, left, top, res_m):
-        with rasterio.open(path, "w", driver="GTiff", width=arr.shape[1], height=arr.shape[0],
-                           count=1, dtype="float32", nodata=-9999.0,
-                           transform=from_origin(left, top, res_m, res_m)) as dst:
-            dst.write(arr.astype(np.float32), 1)
-
-    dc = tempfile.mkdtemp()
-    f = DEEP_COARSEN_FACTOR
-    rng = np.random.default_rng(7)
-    big = (-400.0 + rng.uniform(-40, 40, (8 * f, 8 * f))).astype(np.float32)  # noisy deep field
-    big[:f, :] = -10.0            # one shallow block-row: must stay untouched
-    big[f, 0] = -9999.0           # a nodata pixel inside a deep block: preserved + excluded
-    left0, top0 = -MERC_ORIGIN, MERC_ORIGIN  # window A: block-aligned at the origin
-    _write_dem(f"{dc}/a.tiff", big, left0, top0, 10.0)
-    _deep_coarsen(f"{dc}/a.tiff", factor=f)
-    with rasterio.open(f"{dc}/a.tiff") as src:
-        a = src.read(1)
-    assert np.array_equal(a[:f, :], big[:f, :]), "shallow pixels must be untouched"
-    assert a[f, 0] == -9999.0, "nodata must be preserved"
-    blk = big[f:2 * f, :f]
-    want = blk[blk != -9999.0].mean()
-    got = a[2 * f - 1, f - 1]  # a deep pixel of that block (away from the nodata cell)
-    assert abs(got - want) < 1e-3, f"deep block mean: {got} != {want}"
-    # Window B: the same field cropped at a NON-block-aligned offset (3 px right/down). Values
-    # in the overlap must match window A exactly — the origin-anchored grid, not the window,
-    # defines the blocks. Interior only: A's edge blocks average pad-replicated pixels.
-    off = 3
-    _write_dem(f"{dc}/b.tiff", big[off:, off:], left0 + off * 10.0, top0 - off * 10.0, 10.0)
-    _deep_coarsen(f"{dc}/b.tiff", factor=f)
-    with rasterio.open(f"{dc}/b.tiff") as src:
-        b = src.read(1)
-    assert np.array_equal(a[f:7 * f, f:7 * f], b[f - off:7 * f - off, f - off:7 * f - off]), \
-        "overlapping windows must coarsen identically (seam alignment)"
+    # nodata simplification: a dense OSM-style outline generalizes to the stem's resolution,
+    # shedding vertices while its area barely moves — and the post-clip, per-piece simplify
+    # never pushes a piece's boundary outward across the clip line (the seam contract).
+    from shapely.geometry import box as _box
+    from shapely import get_num_coordinates
+    tol = NODATA_SIMPLIFY_PX * get_resolution(14)
+    ring = [(1000.0 * np.cos(t) + 0.37 * tol * np.sin(60 * t),
+             1000.0 * np.sin(t) + 0.37 * tol * np.cos(60 * t))
+            for t in np.linspace(0, 2 * np.pi, 4000, endpoint=False)]  # dense, sub-tol wobble
+    dense = Polygon(ring + [ring[0]])
+    simp = dense.simplify(tol, preserve_topology=True)
+    assert get_num_coordinates(simp) < get_num_coordinates(dense) / 4, "nodata simplify must shed vertices"
+    assert abs(simp.area - dense.area) < 0.02 * dense.area, "nodata simplify must preserve area"
+    # A piece clipped to a box, then simplified: its vertices are a subset inside the clip box, so
+    # the shared edge stays exactly on the clip line (post-clip simplify never crosses the seam).
+    clipbox = _box(0, -2000, 2000, 2000)  # cuts the disc through its centre at x=0
+    for piece in _polys(dense.intersection(clipbox)):
+        s = piece.simplify(tol, preserve_topology=True)
+        if s.is_empty:
+            continue
+        for sp in _polys(s):
+            xs = [c[0] for c in sp.exterior.coords]
+            assert min(xs) >= -1e-6, "post-clip simplify must not push a vertex across the clip line"
 
     d = tempfile.mkdtemp()
     h = w = 60
