+++ pipelines/smooth.py
@@ -39,6 +39,19 @@ DEPTH_SMOOTH = float(os.environ.get("SMOOTH_DEPTH_SMOOTH", "200")) # ≥ this de
 BLOCK = int(os.environ.get("SMOOTH_BLOCK", "2048"))          # window side (px); caps peak memory
 TRUNCATE = 4.0                                               # gaussian_filter default kernel cutoff (σ)
 
+# Deep-water DEM coarsening: below the threshold, band edges carry no navigational detail, but
+# source noise at fine child_z resolution makes gdal_contour's rings astronomically complex (the
+# deep-basin z14 stems burned 8.5 h+ of pure CPU; docs/plans/2026-07-21-depare-perf.md). Pixels
+# whose 8x-block mean is deeper than the threshold take that mean, so the deep interior contours at
+# ~8x resolution while everything shoaler keeps full detail. -250 m sits between ladder levels (m:
+# -200/-300; ft: -182.88/-365.76), so no level rides the transition. Both the depth-area bands and
+# the contour lines coarsen off this one surface so deep isobaths track the band edges. Only fine
+# windows (child_z >= 12) coarsen — coarse windows are already at or past 8x.
+DEEP_COARSEN_THRESHOLD_M = -250.0
+DEEP_COARSEN_FACTOR = 8
+DEEP_COARSEN_MIN_CHILD_Z = 12
+MERC_ORIGIN = 20037508.342789244  # EPSG:3857 half-extent; block grid anchors here (seam alignment)
+
 
 def halo_px():
     """The gaussian truncation radius (TRUNCATE·σ, +1 for the gradient) in pixels — the window
@@ -111,6 +124,39 @@ def smooth_merged(tmp_folder):
     smooth_tiff(f"{tmp_folder}/{n - 1}-3857.tiff")
 
 
+def deep_coarsen(dem, threshold=DEEP_COARSEN_THRESHOLD_M, factor=DEEP_COARSEN_FACTOR):
+    """Rewrite `dem` in place with deep pixels replaced by their block mean. The block grid is
+    anchored to the EPSG:3857 origin, NOT the window, so overlapping windows (neighbor stems'
+    buffers) average identical blocks and band edges still match at macrotile seams. A pixel is
+    replaced only when it AND its block mean are below the threshold: shelf-edge blocks whose
+    mean dips deep keep their shallow pixels, and a deep canyon inside a shallow block stays
+    fine rather than taking a shallow mean. Nodata pixels are never touched (the nodata pass
+    reads this same raster), and nodata is excluded from block means."""
+    with rasterio.open(dem, "r+") as d:
+        arr = d.read(1)
+        nd = d.nodata
+        res = abs(d.transform.a)
+        h, w = arr.shape
+        # Global block alignment: pad to the origin-anchored grid (pads land in the buffer).
+        col0 = round((d.transform.c + MERC_ORIGIN) / res)
+        row0 = round((MERC_ORIGIN - d.transform.f) / res)
+        padt, padl = row0 % factor, col0 % factor
+        padb, padr = (-(h + padt)) % factor, (-(w + padl)) % factor
+        p = np.pad(arr, ((padt, padb), (padl, padr)), mode="edge")
+        valid = p != nd
+        ph, pw = p.shape
+        vals = np.where(valid, p, 0.0)
+        bsum = vals.reshape(ph // factor, factor, pw // factor, factor).sum(axis=(1, 3))
+        bcnt = valid.reshape(ph // factor, factor, pw // factor, factor).sum(axis=(1, 3))
+        del p, vals, valid
+        bmean = (bsum / np.maximum(bcnt, 1)).astype(np.float32)
+        bmean[bcnt == 0] = nd
+        up = np.repeat(np.repeat(bmean, factor, 0), factor, 1)[padt:padt + h, padl:padl + w]
+        mask = (arr != nd) & (arr <= threshold) & (up != nd) & (up <= threshold)
+        arr[mask] = up[mask]
+        d.write(arr, 1)
+
+
 def _check():
     """Deep flat smooths harder than shallow; shallow stays denoised; a steep step is preserved."""
     rng = np.random.default_rng(0)
@@ -155,6 +201,42 @@ def _check():
     smooth_tiff(i16, block=128)  # must not raise on Int16 (the predictor=3 regression)
     with rasterio.open(i16) as src:
         assert src.read(1).std() < arr16.std(), "Int16 smoothing should denoise"
+
+    # deep_coarsen: shallow and nodata pixels preserved; deep pixels take their block mean;
+    # and the block grid is origin-anchored, so two windows offset against the block grid
+    # produce identical values where they overlap (the macrotile-seam contract).
+    def _write_dem(path, arr, left, top, res_m):
+        with rasterio.open(path, "w", driver="GTiff", width=arr.shape[1], height=arr.shape[0],
+                           count=1, dtype="float32", nodata=-9999.0,
+                           transform=from_origin(left, top, res_m, res_m)) as dst:
+            dst.write(arr.astype(np.float32), 1)
+
+    f = DEEP_COARSEN_FACTOR
+    big = (-400.0 + rng.uniform(-40, 40, (8 * f, 8 * f))).astype(np.float32)  # noisy deep field
+    big[:f, :] = -10.0            # one shallow block-row: must stay untouched
+    big[f, 0] = -9999.0           # a nodata pixel inside a deep block: preserved + excluded
+    left0, top0 = -MERC_ORIGIN, MERC_ORIGIN  # window A: block-aligned at the origin
+    _write_dem(f"{d}/a.tiff", big, left0, top0, 10.0)
+    deep_coarsen(f"{d}/a.tiff", factor=f)
+    with rasterio.open(f"{d}/a.tiff") as src:
+        a = src.read(1)
+    assert np.array_equal(a[:f, :], big[:f, :]), "shallow pixels must be untouched"
+    assert a[f, 0] == -9999.0, "nodata must be preserved"
+    blk = big[f:2 * f, :f]
+    want = blk[blk != -9999.0].mean()
+    got = a[2 * f - 1, f - 1]  # a deep pixel of that block (away from the nodata cell)
+    assert abs(got - want) < 1e-3, f"deep block mean: {got} != {want}"
+    # Window B: the same field cropped at a NON-block-aligned offset (3 px right/down). Values
+    # in the overlap must match window A exactly — the origin-anchored grid, not the window,
+    # defines the blocks. Interior only: A's edge blocks average pad-replicated pixels.
+    off = 3
+    _write_dem(f"{d}/b.tiff", big[off:, off:], left0 + off * 10.0, top0 - off * 10.0, 10.0)
+    deep_coarsen(f"{d}/b.tiff", factor=f)
+    with rasterio.open(f"{d}/b.tiff") as src:
+        b = src.read(1)
+    assert np.array_equal(a[f:7 * f, f:7 * f], b[f - off:7 * f - off, f - off:7 * f - off]), \
+        "overlapping windows must coarsen identically (seam alignment)"
+
     print("smooth.py self-check ok")
